httpd: Moved the golang httpd-logger into the main httpd pacakge
This removes the old ruby logger, and the EM dependency.
