--- httpd-logger/.gitignore
@@ -1,5 +1,0 @@
-debian/files
-debian/tmp
-debian/symbiosis-httpd-logger
-debian/debhelper-build-stamp
-obj-x86_64-linux-gnu/
