--- httpd-logger/debian/changelog
@@ -1,33 +1,0 @@
-symbiosis-httpd-logger (0.3.1) stable; urgency=medium
-
-  * Updated copyright and dh compatibility.
-
- -- Patrick J Cherry <patrick@bytemark.co.uk>  Mon, 24 Apr 2017 15:39:11 +0100
-
-symbiosis-httpd-logger (0.3) stable; urgency=medium
-
-  * Mask out non-permissions bits from file mode inherited from directory
-    (thanks to @banburybill).
-
- -- Patrick J Cherry <patrick@bytemark.co.uk>  Wed, 19 Apr 2017 10:43:19 +0100
-
-symbiosis-httpd-logger (0.2) stable; urgency=medium
-
-  * Non-maintainer upload.
-  * The sync flag is now obeyed (closes #44).
-  * Now creates log files with correct owner (closes #41).
-  * Fixed double logging to the default and per-host logs (closes #45).
-  * Verbose logging normalised.
-  * Debian packaging tidied simplified.
-  * Fixed up dpkg-divert to work properly.
-  * UID/GID for logs is always inherited from the log directory.
-  * Added stricter regex for log lines.
-  * Added prefix flag for testing.
-
- -- Patrick J Cherry <patrick@bytemark.co.uk>  Tue, 22 Nov 2016 12:06:58 +0000
-
-symbiosis-httpd-logger (0.1) unstable; urgency=low
-
-  * Initial release.
-
- -- Steve Kemp <steve@bytemark.co.uk>  Wed, 06 Apr 2016 17:22:17 +0000
