--- httpd-logger/debian/compat
@@ -1,1 +1,0 @@
-9
