--- httpd-logger/debian/control
@@ -1,13 +1,0 @@
-Source: symbiosis-httpd-logger
-Section: web
-Priority: extra
-Maintainer: Steve Kemp <steve@bytemark.co.uk>
-Build-Depends: debhelper (>=9), dh-golang, golang-go (>= 1.5.1)
-Standards-Version: 3.9.4
-
-Package: symbiosis-httpd-logger
-Depends: ${misc:Depends}
-Architecture: any
-Description: golang based httpd logging process for Symbiosis.
- This package replaces the original/default ruby-based logging process
- which is executed by the http daemon, with the intention of saving RAM.
