--- httpd-logger/debian/links
@@ -1,1 +1,0 @@
-usr/bin/symbiosis-httpd-logger usr/sbin/symbiosis-httpd-logger
