--- httpd-logger/debian/rules
@@ -1,13 +1,0 @@
-#!/usr/bin/make -f
-
-export DH_GOPKG := symbiosis-httpd-logger
-
-%:
-	dh $@ --buildsystem=golang --with=golang
-
-#
-# It would be nice to test as part of the build, but that required installing
-# symbiosis-common.
-#
-# override_dh_auto_test:
-#	make test
