--- httpd-logger/debian/source/format
@@ -1,1 +1,0 @@
-3.0 (native)
