+++ httpd/.gitignore
@@ -1,0 +1,2 @@
+obj-x86_64-linux-gnu/
+sbin/symbiosis-httpd-logger
