+++ httpd/Makefile
@@ -1,21 +1,34 @@
 
 RUBYLIB := "${PWD}/lib:${PWD}/../common/lib"
 
-nop:
-	@echo "Valid targets are (alphabetically) :"
-	@echo " "
-	@echo " clean     - Remove bogus files."
-	@echo " docs      - Create manpages for scripts in sbin."
-	@echo " "
+#
+# Not sure this is the correct variable to use, but it appears to give the
+# right name!
+#
+DEB_BUILD_GNU_TYPE := $(shell dpkg-architecture -q DEB_BUILD_GNU_TYPE)
 
-docs: man/symbiosis-httpd-configure.man  man/symbiosis-httpd-logger.man man/symbiosis-httpd-generate-stats.man man/symbiosis-httpd-rotate-logs.man
+all: sbin/symbiosis-httpd-logger docs
 
-test:
+docs: man/symbiosis-httpd-configure.man man/symbiosis-httpd-generate-stats.man man/symbiosis-httpd-rotate-logs.man man/symbiosis-httpd-logger.man
+
+#
+# The dh-golang stuff seems to put the binary in obj-$DEB_BUILD_GNU_TYPE/bin
+#
+obj-$(DEB_BUILD_GNU_TYPE)/bin/%: %.go
+	mkdir -p ${DEB_BUILD_GNU_TYPE}/bin
+	go build -o $@ $<
+
+
+test: obj-$(DEB_BUILD_GNU_TYPE)/bin/symbiosis-httpd-logger
 	$(MAKE) -C vhost-alias test
 	RUBYLIB=${RUBYLIB} ruby test.d/t*.rb
 
-.PHONY: test clean all docs
- 
+sbin/symbiosis-httpd-logger: obj-$(DEB_BUILD_GNU_TYPE)/bin/symbiosis-httpd-logger
+	cp -a $< $@
+
+man/%.txt: ./txt/%.txt
+	cp -a $< $@
+
 man/%.txt: ./sbin/%
 	[ -d man ] || mkdir man
 	RUBYLIB=${RUBYLIB} $< --manual > $@
@@ -26,11 +39,13 @@ man/%.man: man/%.txt
 	test -s $@
 
 clean:
+	$(RM) -r obj-$(DEB_BUILD_GNU_TYPE)
 	cd vhost-alias && make clean
 	@find . -name '.*~' -exec rm \{\} \;
 	@find . -name '.#*' -exec rm \{\} \;
 	@find . -name 'configure-stamp' -exec rm \{\} \;
 	@find . -name 'build-stamp' -exec rm \{\} \;
 	$(RM) -r man
+	$(RM) sbin/symbiosis-httpd-logger
 
-.PHONY: clean nop symbiosis-httpd-man
+.PHONY: test clean all docs
