+++ httpd/apache.d/ssl.template.erb
@@ -120,8 +120,8 @@
         # made against, so that the cron-job in /etc/cron.daily may generate
         # statistics for each domain.
         #
-        ErrorLog   "|| /usr/sbin/symbiosis-httpd-logger -s -u <%= domain.uid %> -g <%= domain.gid %> <%= domain.log_dir %>/ssl_error.log"
-        CustomLog  "|| /usr/sbin/symbiosis-httpd-logger -s -u <%= domain.uid %> -g <%= domain.gid %> <%= domain.log_dir %>/ssl_access.log" combined
+        ErrorLog   "|| /usr/sbin/symbiosis-httpd-logger -s <%= domain.log_dir %>/ssl_error.log"
+        CustomLog  "|| /usr/sbin/symbiosis-httpd-logger -s <%= domain.log_dir %>/ssl_access.log" combined
 </VirtualHost>
 
 <VirtualHost <%= ips.collect{|ip| ip+":80"}.join(" ") %>>
@@ -222,8 +222,8 @@
         # made against, so that the cron-job in /etc/cron.daily may generate
         # statistics for each domain.
         #
-        ErrorLog   "|| /usr/sbin/symbiosis-httpd-logger -s -u <%= domain.uid %> -g <%= domain.gid %> <%= domain.log_dir %>/error.log"
-        CustomLog  "|| /usr/sbin/symbiosis-httpd-logger -s -u <%= domain.uid %> -g <%= domain.gid %> <%= domain.log_dir %>/access.log" combined
+        ErrorLog   "|| /usr/sbin/symbiosis-httpd-logger -s <%= domain.log_dir %>/error.log"
+        CustomLog  "|| /usr/sbin/symbiosis-httpd-logger -s <%= domain.log_dir %>/access.log" combined
 
 % end
 </VirtualHost>
